source/tests/chatbotserver_test/requirements.txt
protobuf~=5.29.3
fastapi~=0.112.2
langgraph~=0.3.9
langchain~=0.3.12
streamlit~=1.42.0
requests~=2.32.3
langchain-core~=0.3.44
langchain-openai~=0.3.8
langchain-community~=0.3.12
langchain-huggingface~=0.1.2
Flask~=3.1.0
langchain-google-genai~=2.1.0
langgraph-checkpoint-sqlite~=2.0.6
uvicorn~=0.32.1
pandas~=2.2.3
openpyxl~=3.1.5
docx2txt~=0.8
pypdf~=4.2.0
chromadb~=0.6.3
pydantic~=2.10.3
pyyaml~=6.0.2
python-dotenv~=0.21.0